requirements.txt
anyio==4.2.0
backoff==2.2.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gql==3.4.1
graphql-core==3.2.3
gspread==6.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
multidict==6.0.4
oauthlib==3.2.2
omnivoreql==0.2.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-telegram-bot==20.7
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.0
StrEnum==0.4.15
typing_extensions==4.9.0
urllib3==1.26.18
yarl==1.9.4
